+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * refactor: Change context array to `Context` object.
 * refactor: Move `Context` to its own namespace.
 * style(Context): Rename `set` to `with` and `remove` to `without`.
+* refactor: Reintroduce `Context` to `MapperInterface`.
 
 ## 0.5.4